requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-active-link==0.1.8
django-crispy-forms==1.14.0
idna==3.3
lxml==4.8.0
Pillow==9.0.1
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8

pandas~=1.4.1
scikit-learn~=1.0.2
cssutils~=2.4.0